Patch BrokerProxy to work around KAFKA-10324.
If we get an empty batch, and are not at the HW, increment the
fetch offset so we don't spin fetching the problematic offset over
and over. Note I had to disable TestBrokerProxy.brokerProxyUpdateLatencyMetrics because it fails with this change. I believe the setup of that test may be invalid; the mock FetchResponse is configured to report the HW as 500, but no messages beyond offset 84 can be returned. Further, the mock SimpleConsumer is configured to report the HW as 100, which means the HW moves backwards for the second part of the test, causing the lag to go negative.